// src/WBObjects/WBNifti.d.ts
declare enum WBNiftiIntent {
    NIFTI_INTENT_NONE = 0,
    NIFTI_INTENT_CORREL = 2,
    NIFTI_INTENT_TTEST = 3,
    NIFTI_INTENT_FTEST = 4,
    NIFTI_INTENT_ZSCORE = 5,
    NIFTI_INTENT_CHISQ = 6,
    NIFTI_INTENT_BETA = 7,
    NIFTI_INTENT_BINOM = 8,
    NIFTI_INTENT_GAMMA = 9,
    NIFTI_INTENT_POISSON = 10,
    NIFTI_INTENT_NORMAL = 11,
    NIFTI_INTENT_FTEST_NONC = 12,
    NIFTI_INTENT_CHISQ_NONC = 13,
    NIFTI_INTENT_LOGISTIC = 14,
    NIFTI_INTENT_LAPLACE = 15,
    NIFTI_INTENT_UNIFORM = 16,
    NIFTI_INTENT_TTEST_NONC = 17,
    NIFTI_INTENT_WEIBULL = 18,
    NIFTI_INTENT_CHI = 19,
    NIFTI_INTENT_INVGAUSS = 20,
    NIFTI_INTENT_EXTVAL = 21,
    NIFTI_INTENT_PVAL = 22,
    NIFTI_INTENT_LOGPVAL = 23,
    NIFTI_INTENT_LOG10PVAL = 24,
    NIFTI_FIRST_STATCODE = 25,
    NIFTI_LAST_STATCODE = 26,
    NIFTI_INTENT_ESTIMATE = 1001,
    NIFTI_INTENT_LABEL = 1002,
    NIFTI_INTENT_NEURONAME = 1003,
    NIFTI_INTENT_GENMATRIX = 1004,
    NIFTI_INTENT_SYMMATRIX = 1005,
    NIFTI_INTENT_DISPVECT = 1006,
    NIFTI_INTENT_VECTOR = 1007,
    NIFTI_INTENT_POINTSET = 1008,
    NIFTI_INTENT_TRIANGLE = 1009,
    NIFTI_INTENT_QUATERNION = 1010,
    NIFTI_INTENT_DIMLESS = 1011
}
declare enum WBNiftiDataType {
    NIFTI_TYPE_UNKNOWN = 0,
    NIFTI_TYPE_UINT8 = 2,
    NIFTI_TYPE_INT16 = 4,
    NIFTI_TYPE_INT32 = 8,
    NIFTI_TYPE_FLOAT32 = 16,
    NIFTI_TYPE_COMPLEX64 = 32,
    NIFTI_TYPE_FLOAT64 = 64,
    NIFTI_TYPE_RGB24 = 128,
    NIFTI_TYPE_INT8 = 256,
    NIFTI_TYPE_UINT16 = 512,
    NIFTI_TYPE_UINT32 = 768,
    NIFTI_TYPE_INT64 = 1024,
    NIFTI_TYPE_UINT64 = 1280,
    NIFTI_TYPE_FLOAT128 = 1536,
    NIFTI_TYPE_COMPLEX128 = 1792,
    NIFTI_TYPE_COMPLEX256 = 2048
}
export { WBNiftiIntent, WBNiftiDataType };